Detect Workspaces During Framework Detection
This commit implements a basic workspace detection as part of framwork
detection. Currently workspace detection only looks at the root
directory but can alter in the future to check child directories if we
identify use cases where that makes sense.

If a workspace is identified we attempt to identify all the associated
packages/paths. We always include two additional directories to check

1. The root directory
2. Docs directory - Based on some very basic research I've found
   multiple cases where a docs directory is used but not included in the
   workspace packages. I think 'docs' is a common enough convention
   where people put docs to incorporate as a default.

One thing I don't like about the current implementation is that we need
to plumb the root directory throughout the detection process so that we
are looking at appropriate paths. I would like to see if there is a
better way to do this.