Simplify Load Settings
This commit removes separate `cfg(test)` and `cfg(not(test))`
implementations of `load_settings`. Previously we had separate
implementations so that in normal execution, not test, we would cache
the loading of settings into a `SETTINGS` singleton object while the
test verion would always load settings from configuration file on disk.

The issue we've run into is that `cfg(test)` isn't visible across crates
so when we execute tests in our Doctavious bin package when it calls
into the cli package we end up calling the `cfg(not(test))`
implementation. As a result the settings we get for the entirety of the
test execution will be whatever was first loaded from one of the tests.

The need for a cached singleton was a premature optimization that I dont
think provides any advantages at this time as we mostly just load
settings once and pass the reference around. At some point it might be
more likely we'll include it in a context object that is passed around
or we'll revisit the need for the singleton.